requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
geopy>=2.4.0
overpy>=0.7
geopandas>=0.14.1
folium>=0.15.0
openai>=1.3.7
python-dotenv>=1.0.0
pydantic>=2.5.0
jinja2>=3.1.2
aiofiles>=23.2.1
numpy>=1.24.3
pandas>=2.0.3
shapely>=2.0.2
scikit-learn>=1.3.2
python-multipart>=0.0.6
streamlit>=1.28.1
plotly>=5.17.0
streamlit-folium>=0.15.0 
google-generativeai>=0.3.0 